+++ bittytax/conv/parsers/__init__.py
@@ -58,5 +58,6 @@ from . import trezorsuite
 from . import uphold
 from . import volt
 from . import wirex
+from . import yoroi
 from . import zelcore
 from . import zerion