+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 - Accounting tool: tax rates and allowance for 2023/24.
 - Exodus parser: added new export format. ([#200](https://github.com/BittyTax/BittyTax/issues/200))
 - Accointing parser: added new export format. ([#219](https://github.com/BittyTax/BittyTax/issues/219))
+- Conversion tool: added parser for Yoroi wallet.
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.